test: Update tests and snapshots
